Move some more panel-specific code out of src/infographic (although they're still fairly intertwined, no avoiding that without a big unnecessary refactor)
