+++ client/src/panels/get_panels_for_subject/get_dept_panels.js
