+++ client/src/panels/get_panels_for_subject/get_gov_panels.js
@@ -27,7 +27,7 @@ import {
   declare_in_year_stat_breakdown_panel,
   declare_gocographic_panel,
   declare_personnel_spend_panel,
-} from '../../panels/panel_declarations/index.js';
+} from '../panel_declarations/index.js';
 
 export const get_gov_panels = subject => ({
   intro: [
